systemc: Create a new method for slave_transactor.
In slave_transactor, a map is added to store coreID and requestorID key-value pairs. When the slave port received request from gem5 sim_controller, it will check the requestorID of pkt and transfer it to different target port according to the map. In systemc world, the multiple sockets can bind to different models. In chi-demo. it will binded to multiple rnf.